% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_clean.R
\name{tf_clean}
\alias{tf_clean}
\title{tf_clean}
\usage{
tf_clean(data)
}
\arguments{
\item{data}{A dataframe to be cleaned}
}
\value{
The cleaned data
}
\description{
Clean TF data
}
\details{
Takes in some specifically formatted data and squares it
}
